requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
efficientnet_pytorch
python-multipart
